Change postfix_form from a vector to a list
I would like to make sure the unary operation (-) is also
taken into account when the postfix expression is evaluated.
In order to accomplish that, I ought to scan the postfix_form
container and cancel consecutive (-)'s. In theory, list has
better performance when it comes to deletion and insertion than
vector. So I changed this container to list to facilitate this
process and possibly make it faster.